mock-chain: refactor object mocking
