+++ packages/mock-chain/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./executor";
 export * from "./mockChain";
 export * from "./mockChainParty";
+export * from "./objectMocking";
